# openapi.yaml
openapi: 3.0.0
info:
  title: FlexCatalog CMS API
  description: API for managing hierarchical elements, schema types, roles, and graph relationships.
  version: 1.0.0
servers:
  - url: /api
paths:
...
  /theme/element/{slug}:
    get:
      summary: Get element data for template rendering (by slug or ID)
      tags: [Templates]
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: Element data with parent, children, and neighbors
  /theme/element/{type}/{id}:
    get:
      summary: Get element data matching the template URL structure
      tags: [Templates]
      parameters:
        - name: type
          in: path
          required: true
          schema: { type: string }
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: Element data with parent, children, and neighbors